Introducing interaction with space walls and a "too close distance"
Introducing a virtual spring as soon as agent gets to close to space
walls just like it does with other agents.

Agent now has a viewable distance that is to be larger than the "too
close distance" at which the agent starts to move away from its
neighbors or the space walls.